Pre-compute the eventual drop_field value in regex parser (#462)
The list of field names is fixed when the regex is compiled. So, we can
compute in advance if the field should be dropped or not, based on if a
regex capture with the same name exists in the pattern. This saves a
branch in the hot path of the transform function.