feat(remap): add variable support (#4802)
This adds variable support to the remap language:

```rust
// variable assignment
$foo = "bar"

// path assignment and variable reference
.bar = $foo
```

This was removed in https://github.com/timberio/vector/pull/4695 (340842453a65875e3898f145bc332d2c4e104e76) to reduce the complexity of that PR and have it move forward without too much unrelated discussions.

Now that all the recent remap-lang changes have been merged, I'm re-introducing the concept of variables in this PR.

## Motivation

I'm happy to _not_ do this if there is opposition to this, but I'll posit my motivation for why I think it makes sense to add variables to the language.

**only manipulate the target object if you need to**

Remap language already supports assigning to paths, but if you want to hold a temporary value, right now you must assign and then delete that value to a field within the target object:

```rust
.url_parts = parse_url(.my_url) // note: we don't support "parse_url(.my_url).scheme"
.scheme = .url_parts.scheme
del(".url_parts")
```

Not only is this inconvenient, it also introduces two failure modes:

* You forget to delete the temporary field, polluting your events
* You assign a temporary value to an existing field, overwriting the previous value

**make conditions more readable**

Now that we support the Remap language in conditions (#4743), variables can help keep those conditions more readable:

```rust
$starts_with_date = match(trim(.timestamp), /^\d{4}-\d{2}-\d{2}.+/)
$other_check = to_bool(/* ... */)
$final_check = /* ... */ || /* ... */

$starts_with_date && $other_check && $final_check
```

**allow variable assignment in immutable contexts**

Not only that, but we plan on introducing an "immutable mode" for Remap (#4744), which would make the target object immutable. This does not apply to variables, since their mutability is "internal", meaning they don't mutate the outside world, and so there is no harm in being able to assign to variables in an immutable context.

Meaning, without variables, you wouldn't even be able to do the things we did in the first and second example because you can't assign values to target paths in an immutable context.

## Syntax

For a discussion and arguments for the `$foo = "bar"` syntax, see the related issue (https://github.com/timberio/vector/issues/4419). The TL;DR is that this syntax is similar to jq, and it prevents accidental variable assignments if you'd write `foo = "bar"` but intended to write `.foo = "bar"` (which would result in an error-free, but unexpected outcome of the script).

## Future Work

With this change, you can **store** values in variables, and **retrieve** those values, but you can't **traverse** containers:

```rust
.foo.bar = "baz" // works
$foo = .foo      // works

$foo             // returns { "bar": "baz" }
$foo.bar         // fails to compile
```

I think the latter should be valid syntax.

closes #4419