fix(influxdb_metrics sink): Fix transmission of counter metrics (#5922)
This reverts commit 4a8f519355efccd456dfff6329d0c0baec2bb5e3.

After reverting this commit, the internal_metrics (and possibly other
metrics sources) ends up sending its metrics through the metrics batch
buffer twice before the batch is flushed to the sink. This allows the
counter handling in the metrics batch buffer to generate the required
incremental record in the batch, which then persists after the batch is
reset. Without it, the metrics are sent once, no record is added to the
batch, and it is never added later.

Signed-off-by: Bruce Guenter <bruce@timber.io>