fix(metrics, sinks): Move normalization out of MetricsBuffer (#6340)
* fix(metrics, sinks): Move metric normalization out of MetricsBuffer

MetricsBuffer included a `state` component that was used to convert
metrics between absolute and incremental kinds. This relied on the batch
buffer never being dropped but rather just regenerated through
`Batch::fresh`. However, `PartitionBatchSink` never calls `fresh` but
rather generates a new batch for each partition and then just drops it.

This moves the state management out of the buffer and into a flat map
before the metrics enter the buffer.

Signed-off-by: Bruce Guenter <bruce@timber.io>