chore: Create skeleton `lib/core` (#7178)
* Create skeleton `lib/core`

Per RFC 7027 we intend to create an extracted vector core. This is the first
step in that RFC's plan of action, to create a skeleton package "core" to
migrate code into. There's very little going on here, just a new library -- with
aggressive clippy settings that we might back out later -- and a correctness and
benchmark test.

Closes #7147

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Add newline to the end of Cargo.toml

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Remove unused benchmark

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>