Make tokio non-optional for core/buffers (#7521)
I incorrectly believed that tokio could be an optional dependency for buffers,
used only when we enabled the disk buffer. This is not accurate. core/buffers
requires tokio in tests. @bruceg reported that the following build string was
broken for him:

> cargo check --workspace --no-default-features --tests

This suggests that our CI has a hole in it.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>