enhancement(aws_kinesis_firehose source): Add record compression support to aws_kinesis_firehose source (#7700)
* Add compression support to aws_kinesis_firehose source

This adds a new configuration option, `record_compression` to the
`aws_kinesis_firehose` source to let users optionall configure the
compression of records within Firehose requests. AWS CloudWatch Logs,
for example, gzips the events before sending them to Firehose.

The default for this is `auto` which uses the newly added `infer` crate
to look at the "magic bytes" from the input to determine if the record
is compressed. Notably, if auto is used, and the record fails to
decompress with the detected encoding, then it is just passed along
as-is since we cannot be sure it isn't a false positive.

However, if the user specifies `gzip` specifically as the
`record_encoding`, then requests will be rejected when the encoding does
not match (the HTTP server will return 400). Firehose will route these
to its dead-letter s3 bucket, if configured.

I attempted to support zstd here too to match the compression support on
the `aws_s3` source but the API of the `zstd` crate is a bit harder to
use (it requires preallocating a sufficient buffer) so I just punted for
now given that gzip is the only common case I'm aware of in Firehose.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* clippy

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>