chore(releasing): Fix commit listing when generating release notes (#7721)
* chore(releasing): Fix commit listing when generating release notes

Currently, when listing commits from the last release during release
note generation, we end up:

* missing commits that appear in master before the last patch release
* listing commits that were released in the last patch release that were
  cherry-picked from master and thus should not be included in the next
  release changelog

Here we add `--right-only --cherry-pick`. `--right-only` limits to
commits that appear in master and excludes commits that are in the patch
release. `--cherry-pick` ensures we supress commits that were
cherry-picked and appear in the last patch release.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>