chore: Teach `Event` what its size is (#8396)
* Teach `Event` what its size is

This commit introduces the `ByteSizeOf` trait into vector-core and implements
this for the `Event` type, focused specifically on the `LogEvent` side of that
structure in testing. This allows us to determine how big `Event` is in memory
without necessarily serializing it. The current implementation for `Event`
trawls through any collections to get the size of the structure but there's no
reason if we need constant time access why this can't be a continually
maintained value.

This has bearing on issue #8263, both in the recognition that the blackhole sink
is slow because of serialization to get byte size and in a similar constraint on
the datadog logs sink for similar reasons. q

Resolves #8272

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* clippy dings

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>