chore: Adjust blackhole sink to use ByteSizeOf trait (#8409)
* Adjust blackhole sink to use ByteSizeOf trait

This commit adjusts the blackhole sink in two important ways. First, it
processes incoming events in a chunk using the `StreamExt::chunk` method,
reducing per-event overhead in this sink to 128th of previous, considering the
chunk size. This does imply that blackhole sink will not remove an event from
the stream until there are 127 of its compatriots waiting, but that seems
acceptable for what blackhole sink is. Second, the sink now uses ByteSizeOf
trait methods to calculate the size of the Event chunk, removing json
serialization cost discussed in #4705.

Blackhole sink is now as fast as it's severely mangled form discussed in #8263.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Tie block size to print_amount, fix rate limit

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>