chore(observability)!: Only refer to components by id rather than name (#8417)
* chore(observability)!: Only refer to components by id rather than name

This PR makes a pass to drop all references to "component name"
preferring "component id" as we believe id is closer to the intent of
the field than name is. Name we expect we may re-add later and have it
be a human readable name.

I'm sure I missed places in the docs and internal variable and type
names, but I believe this PR covers the breaking interface changes:

* Tagging metrics with `component_id` rather than `component_name`
* Removing the `name` field from `Component` in the GraphQL API,
  replacing it with `componentId`

See upgrade the added 0.16.0 upgrade guide.

An alternative approach to this would be to, instead of removing `name`
right away, add `id`. For example, have an additional metric tag of
`component_id` but leave `component_name` too. We could then remove
`component_name` in a subsequent release. I personally feel the impact
should be small enough that users can make the change quickly if needed.
I expect it should just be to update metrics queries in their downstream
systems.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* Fix GraphQL blog post

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* typo

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* PR feedback

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* Re-add component_name as metrics tag

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* More name -> id

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* More name -> id

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* Fix IDs

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* Format

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* Fix vector tap variable shadowing

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>