chore: Introduce a `Batcher` stream processor (#8960)
* Introduce a `Batcher` stream processor

This commit is an extract of the work being done in #8825. The goal
with that work is to change the datadog logs sink to be done purely in
terms of `Stream`, with the goal of improving all our sinks along
similar lines. Most of the code in that PR is hampered by the lack of
streaming batching, whence this commit.

I've introduces a new type into vector-core: `Batcher`. This builds on
the `Partitioner` notion introduced in #8914 to divvy up incoming
items into a key space and ensure that blobs of items are issued when
they consume too much memory, if the blob is full or if the key has
sat for too long unflushed. This commit includes proptest property
tests as opposed to quickcheck with the hope that one day we'll be
able to use
[propverify](https://github.com/project-oak/rust-verification-tools)
in vector.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Adjust type expectations

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Avoid a Poll::Pending

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Avoid shipping under-full batches

Caught by Luke, in the event that the timer fired we would ship underfull
batches for the current partition. Resolved.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Make time async

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* clippy dings

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* comment bug

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* use Delay not Skip

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* documentation request

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* destruct -> into_inner

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Collapse timer response

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* tighten up function

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>