chore: Attempt to enable disallowed-methods clippy lint (#9013)
* Attempt to enable disallowed-methods

This commit enables two new lints in the project: disallowed-methods
and missing-const-for-fn. The first is new to Rust and allows us to
reject some methods from the project if we specify the path fully in
clippy.toml. The second suggests functions that might reasonably be
const, though it has some misfires as the small handful of `allow`
introductions atests to.

disallowed-methods isn't as smooth as I'd like -- namely, I can't get
it to ding introductions of lazy_static -- but it's a start. The const
lint will help vector play better with the optimizer.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* clippy.toml but with newline

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* cargo fmt ding

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* further clippy dings

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>