chore(architecture): add per-batch timeout support to Batcher (#9004)
* initial commit of per-batch timeouts

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* wip commit

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* Implement multi-timer with `KeyedTimer`

This commit builds off the old `Timer` trait and Toby's work to introduce a
multi-timer with property testing. Same approach as before, except now the
testing timer is aware of what keys have been inserted and only responds
"expired" when a key's time comes up. I've also added a standard unit test for
ExpirationQueue -- mostly, see below -- on account of that test needing to
actually check wall-clock. This reasoning here being we know that Batcher will
work with a keyed timer, we have confidence in ExpirationQueue as a keyed timer
so we have confidence in Batcher with ExpirationQueue as its keyed timer.

This more or less works. The property tests pass but the unit test for
ExpirationQueue complains about there being no running reactor. I'm hopeful that
Toby'll know what's up here.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* Re-introduce expiration_map

As Toby pointed out it's not actually possible to re-insert over the top of an
entry and have the timer reset, with regard to DelayQueue. "It's a slab inside."
Gotta have the key to find the spot in the slab, so expiration_map is back in.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* tweak the expirationqueue test to get it in the right shape

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* fix issues with test + docs/fmt/clippy cleanup

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

* fix PR review comments

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>

Co-authored-by: Brian L. Troutwine <brian@troutwine.us>