Slip `ready_chunks` into `Transform::Function` processing (#9228)
This commit introduces a `ready_chunks` into the pipeline for function
transforms. This combinator polls the underlying stream until either it signals
Pending or the chunk limit is reached, sending a vector of items downstream. In
this commit we pull at most 128 `Event` instances off the input stream for
processing, improving locality.

I have not done the same for task transforms, though I think it could be
done. Without changing the trait type we'd need the inverse of `ready_chunks` to
move from a `Stream<Item = Vec<Event>>` into a `Stream<Item = Event>`.

This is worth +7MB/s in #8512

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>