chore: Re-introduce tokio-console under a feature flag (#9240)
* Re-introduce tokio-console under a feature flag

This commit attempts to re-introduce the tokio console to the project,
introduced and removed by @tobz in #8884. If you compile the project with
`RUSTFLAGS="--cfg tokio_unstable"` and the `tokio-console` feature enabled this
does _run_ but nothing shows up in the console yet, which I believe to be a
misconfiguration on our part so far.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Flip on runtime/tokio traces for console

After getting help from @hawkw on discord it turns out the thing the last PR was
missing were targets for `runtime` and `tokio` at trace level in our environment
filter. For now I've set these only to flag on if the console is enabled, as
they're relatively chatty.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>