enhancement(vrl): Add `tally` function (#9015)
* enhancement(vrl): Add `tally` function

Adds a new `tally` function that takes an array of string elements and returns a frequency map of how often each element appears in the list. The naming is modeled after Ruby's `tally` function.

We initially considered a few other options

1. Throw a runtime error if the array has non-string elements in it.
2. Add a mechanism to assert the type of all array elements. This would
  be the same as the first option, but allow us to fail at compile time.
3. Use some serialization format of `vrl::Value` as the keys
4. Update `vrl::Value::Map` to be a `BTreeMap<Value, Value>`. However
  we'd still have to identify how to map this back to `vector::Event`
  which requires `String` keys or error on assignment. We could update
  `vector::Value::Map` to also allow `vector::Value` keys; however, those keys
  would need to be serialized when serializing as, for example, JSON
  however, which only allows String keys.

But decided not to do any of them, but just implement it as a fallible function that only takes arrays of strings so that we could release quickly to unblock a user. We left out the docs to not "release" this function, as we think the implementation may be replaced by iteration support which should be added in Q4.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>