chore: Introduce a 'syslog -> loki' soak (#9729)
Introduce a 'syslog -> loki' soak

This commit does two things. First, it introduces a 'syslog -> loki' soak that
should be helpful in getting #9506 over the line. Second, it introduces
documentation for the process of adding a new soak test and tidies up some of
the terraform.

The new soak can be run like so:

```
> ./soaks/soak.sh syslog_loki 0d1ba5c01d5ae8af3f0d3e12b0e22e528e72901e 12f93fa87c8546d586e790e7ee519989370ef7f0
```

This runs a comparision of the soak between
0d1ba5c01d5ae8af3f0d3e12b0e22e528e72901e (current master as of this commit) and
12f93fa87c8546d586e790e7ee519989370ef7f0 (current tip of #9506 as of this
commit).

Closes #9750

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>