Correct soak service bug, expand observer logging (#10053)
It turns out the previous service definition for vector was overly broad
and allowed for traffic to be routed to pods that were not vector, but
only sometimes. This caused intermitent loss of experimental results.
Observer logging was expanded to help debug this. Credit to Spencer for
the ultimate realization.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Signed-off-by: Spencer Gilbert <spencer.gilbert@datadoghq.com>