Add 'perf' feature flag on regex crate (#11239)
Our grok crate is unusual. We use regex in certain places like date parsing and
onig when dealing with the grok patterns themselves. Work on #11206 strongly
suggests that if sinks are allowed to run less subject to slow receiver issue
then `apply_date_filter` is a CPU hotspot. As suggested by @tobz we enable the
'perf' flags to the regex crate, which features heavily in the implementation of
that function.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>