Correct detect_regressions drift filter bug (#11273)
In the analysis script we are interested in the absolute magnitude of any mean
swing. That is not the case in detect_regressions, where we only care for
negative movements. The logic was inappropriately moved from the one script to
the other, now fixed. Regressions will no longer be flagged for _positive_
changes, only negative.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>