Remove dynamic dispatch in Conditions (#11349)
This commit removes the notion of `Box<dyn Condition>` from the code base, using
a Big Old Enum approach instead. This is suggested by the work on #10144 where
Condition always ranks highly in on-cpu/off-cpu measures. It is hoped but not
knowing that removing dynamic dispatch in this hotspot will have a measurable
bump on condition heavy configurations, of which pipeline is a notable
instance.

This is roughly similar to the approach seen in #11200 but in a different area.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>