Add a global deny unreachable_pub, with carve outs (#11538)
This commit makes `deny(unreachable_pub)` the default in the project, allowing
for carve-outs for large sections of the tree. This is consistent with my
background work recently to reduce the exposed interface of vector. I now have
specific areas to aim retypist at and we scope growth of the problem to only a
subset of the tree.

This flag does not apply to any sub-crates.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>