chore: Remove the disabled http-to-http-noack soak (#11717)
Remove the disabled http-to-http-noack soak

It turns out the behavior of the runner has shifted over time. Used to be we
could put a soak under `disabled/` and the test would not run. This is no longer
the case and the previous PR -- #11712 -- auto-merged despite not having its
required checks, breaking master.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>