chore: Update lading, remove minikube from the soaks (#11891)
chore: Update lading, remove minikube from the soaks

This commit updates lading, consistent with the work done in
https://github.com/blt/lading/pull/148. This removes the need for this
project to run a minikube for the purposes of soak testing, with
bearing on #11772. It is hoped but not demonstrated that with minikube
gone we'll have less inexplicable behavior from the soak tests. We now
are able to capture stdout/stderr logs as artifacts in the CI process,
which should aid in debugging. The lading update also allows for
minikube, as we no longer have to stand up and tear down a kube.

With more logic pushed into lading the complexity load on this project
shrinks substantially. It _should_ also be possible for folks on OS X
and for users of btrfs to run the soaks locally now. Because the
process of collecting samples has improved we are now also running all
replicas on one instance, back to back. Using the time savings we also
run more replicas, 6 instead of 3. This should improve Github
responsiveness -- one artifact bundle per soak, plus plots and images
-- and also give us more samples to make regression arguments from.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>