enhancement(aws_sqs source): Use single task for finalization (#12138)
The aws_sqs source was spawning a tokio task for each batch of input
messages in order to asynchronously delete acknowledged messages. This
moves it to use the new `UnorderedFinalizer` framework to handle that in
a single task. Along the way, that framework is modified to run the
"done" task asynchronously.