chore(ci): Run check-component-features in parallel (#12465)
This speeds up the `check-component-features` check by running each
feature in a separate job in a separate build tree. This is possible
because cargo doesn't overwrite output files after they are created. To
take advantage of this, we first do a check on the full build and then
create a temporary build tree for each feature composed of hard links to
the individual files created during that check. This speeds up the
subsequent builds and allows them to run in parallel as they no longer
clobber each other's outputs.