enhancement(geoip transform): Bring up to parity with Logstash geoip filter (#12803)
* enhancement(geoip transform): Bring up to parity with Logstash geoip filter

This brings the `geoip` tranfsorm to parity with what the logstash geoip
filter returns. Note that the data doesn't match the same shape, but is
convertable to logstash's schema by mapping fields VRL.

There is a bunch more in the database we could enrich here, but I stuck
with what logstash currently supports since this was the user need.

I tried to withold rewriting the world here, but noticed some othe
oddities with this transform which is probably due to its age:

* It always inserts all fields into the log event (typically as empty
  strings) even if the fields aren't returned by the database. I think
  ideally it would either: not insert fields it couldn't find, or insert
  them with `null` values as I did here for the added `metro_code`
* `latitude` and longitude` are inserted as strings despite being f64s
* The transform builds up a temporary struct that is then converted into
  a JSON value to be inserted into the log event. I think just using
  a BTreeMap here would make more sense.

We have an issue to integrate this with VRL
(https://github.com/vectordotdev/vector/issues/9283). I think we should
avoid these issues there where since it is reasonable to make
a backwards incompatible change as users will need to opt in to cut over
to using it.

Ref: https://github.com/logstash-plugins/logstash-filter-geoip

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>