fix(loki sink): Fix all uses of `Write::write` (#12873)
This sweeps the sources for uses of the base `write` method that don't
account for short writes. Some uses are part of implementing `trait
Write`, so the callers would be accounting for short writes and so don't
need fixing.  This only showed up as a possible issue for the `loki`
sink and the core config generate code.