chore(ci): Fix application of `rustflags` (#13074)
* chore: Fix overwriting cargo config in bootstrap

The Ubuntu bootstrap script overwrites the `.cargo/config.toml` file in
order to install the `mold` linker. This leads to the removal of the
clippy flag options that provide extra links that we want to apply to
all crates. This was discovered by `make check-clippy` passing in CI but
failing on non-CI systems.

This change appends to the cargo config instead of overwriting it to
allow both.

* chore: Fix application of rustflags on Ubuntu

It turns out setting `$RUSTFLAGS` in the GitHub workflow file overrides
all the settings in `.cargo/config.toml` anyways, so add a guard against
setting that in the bootstrap and move the settings there.

Co-authored-by: Jesse Szwedko <jesse@szwedko.me>