fix(observability): Lower log level for HTTP errors (#12930)
* fix(observability): Lower log level for HTTP errors

This lowers the log level for errors seen at the HTTP client level to
debug. If the error is fatal and won't be retried, it will be logged at
a higher level by the component using the HTTP client. As it stands, it
is confusing to users that they see an error log for a request failure
that will automatically be retried.

It's possible we shouldn't be emitting the metrics at this level either,
but I'll leave that as a follow-up as part of implementing
https://github.com/vectordotdev/vector/pull/12822.

Closes: #12850

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>

* Update src/internal_events/http_client.rs