chore: Move TLS library to `vector-core` (#13913)
This reduces the knot of feature flag dependencies for a relatively small
but core library.

Basically, I was trying to run one of the unit tests with minimal flags, but gave up trying to figure out all of the needed "utils" flags I needed to run the test. This is the fruit of that frustration as it removes a number of those flags.