revert(observability): Convert internal metric counts back to absolute (#14251)
Revert "enhancement(observability): Convert internal metric counters to incremental (#13883)"

This reverts commit 9a14e9ce42a529a6319c60c98ac12887b8d4b991.

The above change broke metrics exposed in `vector top` and the API, as they
assumed the values were absolute. Also, the current setup does not work in the
presence of multiple consumers pulling from the metrics registry, as the
increments will be split across the consumers with none seeing the whole value.