fix(prometheus_exporter sink): Avoid extra work in bad requests (#14869)
As written, the cost of counting event byte size is taken even if the request is
unauthorized or hits the wrong path. This scopes the count operation down into
the path where the request was successful. As a side effect, it also prevents
the sink from emitting the events sent counter when no events were sent.