feat(ci): Introduce 'regression detection' flow (#14935)
This pull request introduces the 'regression detection' flow, a more advanced version of the existing 'soak' flow. The basic idea is the same: use a carefully controlled experimental setup to make a statistical argument about the performance impact of small changes. The regression detector has a proper team behind it now -- Single Machine Performance -- and has a more effective statistical model behind it.

This PR lacks analysis. This is expected. SMP does not yet have analysis working in our private compute environment -- please chat with @goxberry for details -- but we will soon. That said, it's very helpful for us to have the regression detector running on live PRs to understand if we've got your cluster right-sized etc.

We will add analysis via PR in the future. For now, the regression detector runs in parallel to the existing soaks but does not compete with them for compute resources in any way.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>