enhancement(elasticsearch sink): Reintroduce `api_version` option  (#15082)
Reintroduces #14918 that was reverted in #15006. It turned out that this is a breaking change so it was decided to not abort Vector on failure to detect Elasticsearch version. Instead we log a warning and estimate the version. This will make the transition smoother. In a future version this can be changed to fatal error.