fix(ci): Correct Regression Detector variant checkout (#15289)
* fix(ci): Correct Regression Detector variant checkout

This PR is intended to correct the problem identified in #15230, namely that the
Regression Detector is not correctly building baseline variant separately from
comparison variant and is instead building two copies of the comparison. This is
a consequence of Github Actions' willingness to treat a non-existent 'need' as
an empty string and actions/checkout's willingness to treat an empty string as
non-existent field.

Quite annoying to debug, relatively simple to fix. We have confirmed the
efficacy of this fix in #15287 where we intentionally introduced a break into
the comparison side. Note now that in that PR baseline builds as expected,
comparison does not.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* missed rename

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>