chore(config): Move the `SchemaGenerator` into a `RefCell` (#16538)
Due to a [bug](https://github.com/rust-lang/rust/issues/57349#issuecomment-1266769993)
in the implementation of const functions, we cannot create references to
functions that have `mut` references in the parameter lists. Since the
`generate-schema` function currently takes such a reference, this wraps the
value in a `RefCell` and handles mutability within `get_or_generate_schema`,
which is the only location that requires it.