chore(config): Simplify log schema init (#17012)
The top-level `init_log_schema` function would load and parse the given config
file paths and then use the `log_schema` data from that to setup the initial
data. For both the run and validate commands, however, the immediate next step
is to load the same config files again. This change moves the log schema init
after this load step and uses the log schema from there to initialize.