chore: update `vrl` to `0.4.0` (#17378)
This upgrades VRL to `0.4.0`


Notable changes:
- This is the first crates.io release for VRL. It's no longer a git
dependency!
- All VRL macros are now exported at the root, which required some
import changes
- Previously the `vrl` crate had an internal `test` feature that was
un-intentionally enabled. This was caught and fixed in recent VRL
refactoring. Vector was relying on this in a few places where is
shouldn't be (all related to converting an `f64` into a `Value`. That
implementation is normally only available for tests, since the `f64`
needs to be checked for `NaN` first). As a quick fix, to keep existing
behavior, the `test` feature is now explicitly enabled for VRL. [An
issue](https://github.com/vectordotdev/vector/issues/17377) was created
to track this and remove it.