chore(statsd sink): refactor `statsd` sink to stream-based style (#16199)
This PR completely refactors the `statsd` sink in the "new style", which
uses stream-based combinators to build a `Stream` implementation that
drives component behavior.

At a high-level, the PR is indeed for refactoring the sink, but
ultimately includes as much, if not more, refactoring work around
establishing reusable `Service`-based primitives for building other
sinks like `statsd` i.e. `socket` or `syslog`.

## Reviewer Notes

I've mostly copied the existing shared socket sink types --
`TcpSinkConfig`, etc -- and existing socket services -- `UdpService` --
and created consistent versions of them for TCP, UDP, and Unix Domain
sockets.

This includes a configuration type that is `Configurable`-compatible for
all of them, with socket-specific configurations[1] and then methods for
generating both the `Service` implementation and a `Healthcheck`
implementation. Ultimately, this should form the basis of other sink
refactors that use sockets directly (`socket`, `syslog`, etc) but it may
be desirable to do some more trait-ifying to avoid some of the necessary
boilerplate introduced here.

## Remaining Work

- [x] fix normalizer unit tests + add one for sketches