chore: add sink prelude (#17595)
Generally our stream based sinks will share a lot of imports. This adds
a `crate::sinks::prelude` module that can be imported to bring in the
commonly used imports. The advantage is not only reducing the size of
our imports and making it easier to maintain (often changes to the
framework result in having to change the imports of all the sinks
modules which is tedious) but it also encourages and guides using the
shared components available.

This also updates the following sinks to use the prelude:

- amqp
- kafka
- aws_kinesis
- loki
- pulsar

---------

Signed-off-by: Stephen Wakely <fungus.humungus@gmail.com>