chore(config): Make config schema output ordered (#17694)
Key/value maps in the config schema currently use an `IndexMap`. This
provides for stable ordering of the output between runs. Unfortunately,
due to the use of initializers via `inventory`, that ordering is
dependent on the order in which those initializers are run, which is in
turn dependent on the choices the linker (and link-time optimization)
makes when building Vector. The result is that config schemas can be
very hard to compare between builds due to large blocks of it moving
around in the output file.

This change changes this type to a `BTreeMap`, which is fully ordered
internally. This will result in a consistent ordering of maps, hopefully
producing config schema output that is more consistent.

<!--
**Your PR title must conform to the conventional commit spec!**

  <type>(<scope>)!: <description>

  * `type` = chore, enhancement, feat, fix, docs
  * `!` = OPTIONAL: signals a breaking change
* `scope` = Optional when `type` is "chore" or "docs", available scopes
https://github.com/vectordotdev/vector/blob/master/.github/semantic.yml#L20
  * `description` = short description of the change

Examples:

  * enhancement(file source): Add `sort` option to sort discovered files
  * feat(new source): Initial `statsd` source
  * fix(file source): Fix a bug discovering new files
  * chore(external docs): Clarify `batch_size` option
-->