chore(observability): emit `component_sent` events by `source` and `service` (#17549)
Closes #17580 
Closes #17581 

This is still in draft until I can get the following done.

- [ ] ~~There are way more clones that I am happy with here, especially
since this is in a hot path. These need reducing.~~

The remaining clones that I would like to remove are in the `get_tags`
functions. This didn't seem trivial, and given the fairly positive
regression numbers, I think it should be ok to defer for now.

- [x] Function documentation.
- [ ] Currently source schemas aren't being attached to the event at
runtime, so the service meaning can't be retrieved. That won't work
until this has been done. This will be a separate PR - #17692
- [x] I've only tested this with the kafka sink so far. I think it
should work with all Stream sinks without needing any further
modification - but further testing is needed.
- [x] Tests. A bunch of tests need writing. 
- [x] The Vector source tests are failing I think because we now have
`EventsSent` and `TaggedEventsSent` which both emit
`component_sent_event` events and the test framework doesn't like this.
This needs fixed.
- [ ] We will need to review every sink to ensure they work with this.
All the stream based sinks should, but the others are highly likely to
need some work.

---------

Signed-off-by: Stephen Wakely <fungus.humungus@gmail.com>