chore(ci): fix gardener issues comment workflow (#17825)
As it was previously implemented, this workflow would add an issue to
the Gardener board if it wasn't already present. This updates the
graphql query to fetch the necessary project item id without creating
new items.

Tested the graphql query in the
[playground](https://docs.github.com/en/graphql/overview/explorer).
Tested the `jq` command locally. 
Example Issue ID: `I_kwDOCLjIc85p9V5j`

---------

Co-authored-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>