chore(CI): Switch regression detector to new API and analysis service (#17912)
This should be a transparent change. If you notice any issues, please
ping us in #single-machine-performance. We'll also be keeping a close
eye on things after this is merged.

## Single Machine Performance API (`curta`)

The Single Machine Performance team has built an API to interact with
the regression detector system. This will allow us to maintain
reliability while continuing to evolve the system. This is enabled with
the `--use-curta` flag, which will become the default in an upcoming
release.

The curta API uses a unique URI for each team. This isn't secret - all
endpoints have authorization - but we'd prefer it not be public. The URI
for Vector is in SMP-617. The `smp` binary can read this from the
`SINGLE_MACHINE_PERFORMANCE_API` environment variable or the
`--api-base` flag.

## Lambda Analysis Worker (`consignor`)

We have also changed how the analysis worker is deployed so it can run
on-demand for incoming requests. This improves analysis speed while
reducing our operational workload. This is set using the
`--use-consignor` flag, which will also become the default in a future
release.

---------

Co-authored-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>