chore(observability): consolidate `EventCountTags` with `TaggedEventsSent` (#17865)
@bruceg pointed out that the structs `EventCountTags` and
`TaggedEventsSent` are basically the same, so there is no need for them
to be separate structs.

This consolidates the two structs. The main change has been updating the
`registered_event` macro to derive `Clone, PartialEq, Eq, PartialOrd,
Ord, Hash` to the event structs. This hasn't caused anything to break,
but maybe it could cause problems down the line? `PartialOrd` and `Ord`
weren't strictly needed, but they are used in a test to sort the events
to ensure the test is deterministic.

---------

Signed-off-by: Stephen Wakely <fungus.humungus@gmail.com>