fix(dev): fix issues when using container tools and `cargo` is not installed locally (#18112)
* fix(dev): fix error in `make` when `cargo` is not installed

Fixes the following error when using `make environment`:

    make: cargo: No such file or directory

Signed-off-by: Hugo Hromic <hhromic@gmail.com>

* fix(dev): fix `build` make target when `cargo` is not installed

Fixes the following error when using `make build ENVIRONMENT=true`:

    Makefile:212: *** "Please install Rust: https://www.rust-lang.org/tools/install".  Stop.

Also switch to using `command` (POSIX compatible) instead of `which` for detecting `cargo`.
Details in <https://mywiki.wooledge.org/BashFAQ/081>.

Signed-off-by: Hugo Hromic <hhromic@gmail.com>

* fix(dev): fix detection of container tool in `make`

Fixes the following error when `ENVIRONMENT=true` and neither Docker or Podman are available:

    make: podman: No such file or directory
    make: *** [Makefile:174: build] Error 127

Podman was being used as fallback even if also not available.
A descriptive error is now reported instead in this situation.

Signed-off-by: Hugo Hromic <hhromic@gmail.com>

* chore: add `ifneq` Makefile keyword to spell checker allow-list

Signed-off-by: Hugo Hromic <hhromic@gmail.com>

---------

Signed-off-by: Hugo Hromic <hhromic@gmail.com>