feat(route transform): Add option to enable/disable unmatched output (#18309)
* feat(route transform): Add option to enable/disable unmatched output

This commit adds a new boolean option `reroute_unmatched` to the `route` transform.

It is inspired on the `reroute_dropped` option in the `remap` transform, allowing the user to control
if they want or not the `<transform_name>._unmatched` output to be created and used. For backwards
compatibility, this new option defaults to `true`.

Users that are not interested in processing unmatched events can use this option to avoid the following
warning on Vector startup, and also to remove the unnecessary `_unmatched` output in `vector top`:
```
WARN vector::config::loading: Transform "route._unmatched" has no consumers
```

Signed-off-by: Hugo Hromic <hhromic@gmail.com>

* Update existing `can_serialize_remap` test

* Add test for the new `reroute_unmatched` option

---------

Signed-off-by: Hugo Hromic <hhromic@gmail.com>