chore(dev): Add DEPRECATIONS.md file to track deprecations (#18613)
* chore(dev): Add DEPRECATIONS.md file to track deprecations

Add a file to track deprecations through their lifecycle and remind us to come back to take any
deprecation actions as part of each release.

See https://github.com/vectordotdev/vector/issues/14983 for the process proposal. The intent is to
make sure we follow through on deprecations, migrations, and removals.

Closes: https://github.com/vectordotdev/vector/issues/14983

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>

* Update process in DEPRECATION.md too

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>

* Remove from "to be removed"

Will instead be added once the migration step is complete

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>

* Fix markdown headers and add link to process

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>

---------

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>