chore(core): Remove optionality from topology controller reload (#20010)
The `TopologyController::reload` function took a configuration parameter to load
as `Option<Config>`. However, if the config was `None`, it would immediately
return with `ReloadOutcome::NoConfig`, which was the also only place that return
value could be generated. Instead, this moves the handling of missing
configurations to the two callers, simplifying the reload semantics.